Avoid circular dependency for user form
